+++ src/Libplanet.Action/ActionContext.cs
@@ -1,9 +1,10 @@
-using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Threading;
+using System.Diagnostics.Contracts;
 using Libplanet.Action.State;
 using Libplanet.Crypto;
+using Libplanet.Types.Assets;
+using Libplanet.Types.Blocks;
 using Libplanet.Types.Evidence;
 using Libplanet.Types.Tx;
 
@@ -11,10 +12,6 @@ namespace Libplanet.Action
 {
     internal class ActionContext : IActionContext
     {
-        public static readonly AsyncLocal<GasMeter> GetGasMeter = new AsyncLocal<GasMeter>();
-
-        private readonly long _gasLimit;
-
         private readonly IReadOnlyList<ITransaction> _txs;
 
         public ActionContext(
@@ -23,10 +20,11 @@ namespace Libplanet.Action
             Address miner,
             long blockIndex,
             int blockProtocolVersion,
+            BlockCommit? lastCommit,
             IWorld previousState,
             int randomSeed,
             bool isPolicyAction,
-            long gasLimit,
+            FungibleAssetValue? maxGasPrice,
             IReadOnlyList<ITransaction>? txs = null,
             IReadOnlyList<EvidenceBase>? evidence = null)
         {
@@ -35,14 +33,13 @@ namespace Libplanet.Action
             Miner = miner;
             BlockIndex = blockIndex;
             BlockProtocolVersion = blockProtocolVersion;
+            LastCommit = lastCommit;
             PreviousState = previousState;
             RandomSeed = randomSeed;
             IsPolicyAction = isPolicyAction;
-            _gasLimit = gasLimit;
+            MaxGasPrice = maxGasPrice;
             _txs = txs ?? ImmutableList<Transaction>.Empty;
             Evidence = evidence ?? ImmutableList<EvidenceBase>.Empty;
-
-            GetGasMeter.Value = new GasMeter(_gasLimit);
         }
 
         /// <inheritdoc cref="IActionContext.Signer"/>
@@ -60,6 +57,9 @@ namespace Libplanet.Action
         /// <inheritdoc cref="IActionContext.BlockProtocolVersion"/>
         public int BlockProtocolVersion { get; }
 
+        /// <inheritdoc cref="IActionContext.LastCommit"/>
+        public BlockCommit? LastCommit { get;  }
+
         /// <inheritdoc cref="IActionContext.PreviousState"/>
         public IWorld PreviousState { get; }
 
@@ -69,6 +69,10 @@ namespace Libplanet.Action
         /// <inheritdoc cref="IActionContext.IsPolicyAction"/>
         public bool IsPolicyAction { get; }
 
+        /// <inheritdoc cref="IActionContext.MaxGasPrice"/>
+        [Pure]
+        public FungibleAssetValue? MaxGasPrice { get; }
+
         /// <inheritdoc cref="IActionContext.Txs"/>
         public IReadOnlyList<ITransaction> Txs => IsPolicyAction
             ? _txs
@@ -77,40 +81,7 @@ namespace Libplanet.Action
         /// <inheritdoc cref="IActionContext.Evidence"/>
         public IReadOnlyList<EvidenceBase> Evidence { get; }
 
-        /// <inheritdoc cref="IActionContext.UseGas(long)"/>
-        public void UseGas(long gas) => GetGasMeter.Value?.UseGas(gas);
-
         /// <inheritdoc cref="IActionContext.GetRandom"/>
         public IRandom GetRandom() => new Random(RandomSeed);
-
-        /// <summary>
-        /// Returns the elapsed gas of the current action.
-        /// </summary>
-        /// <returns>
-        /// The elapsed gas of the current action.
-        /// </returns>
-        /// <exception cref="InvalidOperationException">
-        /// Thrown when <see cref="GetGasMeter"/> is not initialized.
-        /// </exception>
-        public long GasUsed()
-        {
-            if (GetGasMeter.Value is { } gasMeter)
-            {
-                return gasMeter.GasUsed;
-            }
-
-            throw new InvalidOperationException($"{nameof(GetGasMeter)} is not initialized.");
-        }
-
-        /// <summary>
-        /// Returns the gas limit of the current action.
-        /// </summary>
-        /// <returns>
-        /// The gas limit of the current action.
-        /// </returns>
-        public long GasLimit()
-        {
-            return _gasLimit;
-        }
     }
 }