+++ test/Libplanet.Action.Tests/ActionEvaluationTest.cs
@@ -1,9 +1,12 @@
+using System.Collections.Immutable;
+using System.Numerics;
 using Bencodex.Types;
 using Libplanet.Action.State;
 using Libplanet.Action.Tests.Common;
 using Libplanet.Crypto;
 using Libplanet.Mocks;
 using Libplanet.Types.Blocks;
+using Libplanet.Types.Consensus;
 using Serilog;
 using Xunit;
 using Xunit.Abstractions;
@@ -27,8 +30,26 @@ namespace Libplanet.Action.Tests
         [Fact]
         public void Constructor()
         {
-            var txid = new System.Random().NextTxId();
+            var random = new System.Random();
+            var txid = random.NextTxId();
             Address address = new PrivateKey().Address;
+            var key = new PrivateKey();
+            var hash = random.NextBlockHash();
+            var lastCommit = new BlockCommit(
+                0,
+                0,
+                hash,
+                new[]
+                {
+                    new VoteMetadata(
+                        0,
+                        0,
+                        hash,
+                        DateTimeOffset.UtcNow,
+                        key.PublicKey,
+                        BigInteger.One,
+                        VoteFlag.PreCommit).Sign(key),
+                }.ToImmutableArray());
             IWorld world = new World(MockWorldState.CreateModern());
             world = world.SetAccount(
                 ReservedAddresses.LegacyAccount,
@@ -41,10 +62,11 @@ namespace Libplanet.Action.Tests
                     address,
                     1,
                     Block.CurrentProtocolVersion,
+                    lastCommit,
                     new World(MockWorldState.CreateModern()),
                     123,
                     false,
-                    0),
+                    maxGasPrice: null),
                 world);
             var action = (DumbAction)evaluation.Action;
 
@@ -62,6 +84,7 @@ namespace Libplanet.Action.Tests
                 (Text)"item",
                 evaluation.OutputState.GetAccount(ReservedAddresses.LegacyAccount).GetState(address)
             );
+            Assert.Equal(lastCommit, evaluation.InputContext.LastCommit);
         }
     }
 }