+++ test/Libplanet.Net.Tests/SwarmTest.cs
@@ -1284,7 +1284,7 @@ namespace Libplanet.Net.Tests
             for (int i = 0; i < 6; i++)
             {
                 Block block = chain.ProposeBlock(
-                    ChainPrivateKey, TestUtils.CreateBlockCommit(chain.Tip));
+                    GenesisProposer, TestUtils.CreateBlockCommit(chain.Tip));
                 chain.Append(block, TestUtils.CreateBlockCommit(block));
             }
 
@@ -1324,7 +1324,7 @@ namespace Libplanet.Net.Tests
             for (int i = 0; i < 6; i++)
             {
                 Block block = chain.ProposeBlock(
-                    ChainPrivateKey, TestUtils.CreateBlockCommit(chain.Tip));
+                    GenesisProposer, TestUtils.CreateBlockCommit(chain.Tip));
                 chain.Append(block, TestUtils.CreateBlockCommit(block));
             }
 
@@ -1365,7 +1365,7 @@ namespace Libplanet.Net.Tests
             for (int i = 0; i < 6; i++)
             {
                 Block block = chain.ProposeBlock(
-                    ChainPrivateKey, CreateBlockCommit(chain.Tip));
+                    GenesisProposer, CreateBlockCommit(chain.Tip));
                 chain.Append(block, TestUtils.CreateBlockCommit(block));
             }