+++ test/Libplanet.Tests/Action/WorldTest.cs
@@ -141,10 +141,11 @@ namespace Libplanet.Tests.Action
                 signer,
                 0,
                 ProtocolVersion,
+                null,
                 world,
                 0,
                 false,
-                0);
+                null);
         }
 
         [Fact]