+++ test/Libplanet.Tests/Consensus/ValidatorSetTest.cs
@@ -167,18 +167,18 @@ namespace Libplanet.Tests.Consensus
             var blockCommitWithUnorderedVotes =
                 new BlockCommit(height, round, hash, unorderedVotes);
             var blockCommitWithInvalidPowerVotes =
-                new BlockCommit(height, round, hash, unorderedVotes);
+                new BlockCommit(height, round, hash, invalidPowerVotes);
             var blockCommitWithInsufficientVotes =
                 new BlockCommit(height, round, hash, orderedVotes.Take(5).ToImmutableArray());
             var validBlockCommit = new BlockCommit(height, round, hash, orderedVotes);
 
-            Assert.False(
+            Assert.Throws<InvalidBlockCommitException>(() =>
                 validatorSet.ValidateBlockCommitValidators(blockCommitWithUnorderedVotes));
-            Assert.False(
+            Assert.Throws<InvalidBlockCommitException>(() =>
                 validatorSet.ValidateBlockCommitValidators(blockCommitWithInvalidPowerVotes));
-            Assert.False(
+            Assert.Throws<InvalidBlockCommitException>(() =>
                 validatorSet.ValidateBlockCommitValidators(blockCommitWithInsufficientVotes));
-            Assert.True(validatorSet.ValidateBlockCommitValidators(validBlockCommit));
+            validatorSet.ValidateBlockCommitValidators(validBlockCommit);
         }
     }
 }